pyproject.toml
[project]
name = "databricks-governance"
version = "1.0.0"
description = "Production Governance and MLOps with Databricks — hands-on labs for Unity Catalog, CI/CD, monitoring, and secrets"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0"}

dependencies = [
    "pyspark>=3.5.0",
    "delta-spark>=3.0.0",
    "mlflow>=2.10.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.4.0",
    "rich>=13.0.0",
    "databricks-sdk>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["examples"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "E722"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = ["venv", ".venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
